Use python2 or python3 explicitly as the interpreter depending on when PY3 or PY3ONLY is used as the version.
This is needed because /usr/bin/python is an ambigous symlink on most
Linux distrubtions, some distro links it to python2.7, some distro
likes gentoo may link it to python3. Given python2 and python3 are not
fully compatible, we should choose the interpreter explicitly based on
the version specified in py_binary rule.

This should fix issue #200

Change-Id: I6f62b080b795112b1945ada5bb2ad7e58f956c2c